Split the plugins.js to multiple small files
